- Move StateBrowserUsermenu into own component - Place StateBrowserUsermenu at top of statebrowser - Move styling and functions over to Statebrowserusermenu
